Color Wheel Update
Now, you can change the color of the pen. By default, the color is black and there are premade colors of black, red, and blue that are always there. But there is also a color palette button that when selected opens up a color wheel where you can select any color.